+++ .gitignore
@@ -1,2 +1,2 @@
-rosbags/
+rosbags/*
 UserConfig.txt
