+++ AUTHORS
@@ -1,0 +1,4 @@
+Rafael Pérez Seguí
+Miguel Fernández Cortizas
+Pedro Arias Pérez
+David Pérez Saura
