+++ README.md
@@ -1,1 +1,1 @@
-#  project_dji_m210_basic_mission
+#  project_dji_osdk
