+++ real_config/platform_config.yaml
@@ -1,0 +1,4 @@
+/**:
+  ros__parameters:
+    cmd_freq: 60.0  # Hz of platform commands send
+    info_freq: 10.0  # Hz of platform info publish
