+++ real_config/state_estimator_odom.yaml
@@ -1,0 +1,5 @@
+/**:
+  ros__parameters:
+    odometry_topic: "sensor_measurements/odom"
+    use_gps: false
+    set_origin_on_start: false
