+++ rosbag/record_rosbag.bash
@@ -1,32 +1,32 @@
 #!/bin/bash
 
-drone_namespace=$1
-drone_namespace=${drone_namespace:="drone0"}
+source utils/tools.bash
 
-echo $(pwd)
+# Get drone namespaces from command-line argument
+drone_namespaces_string=$1
+drone_namespaces=($(string_to_list "$drone_namespaces_string"))
 
-reliability_override_file=$(pwd)/rosbag/reliability_override.yaml
-if [[ "$drone_namespace" == "M200" ]] 
-then
-    reliability_override_file=$(pwd)/rosbag/reliability_override_M200.yaml
-elif [[ "$drone_namespace" == "M300" ]] 
-then
-    reliability_override_file=$(pwd)/rosbag/reliability_override_M300.yaml
-fi
+echo "Recording rosbag for drones: ${drone_namespaces[@]}"
 
+# Create directory for rosbags
 mkdir rosbag/rosbags 2>/dev/null
-cd rosbag/rosbags &&\
-ros2 bag record \
-"/$drone_namespace/platform/info" \
-"/$drone_namespace/self_localization/pose" \
-"/$drone_namespace/self_localization/twist" \
-"/$drone_namespace/actuator_command/twist" \
-"/$drone_namespace/sensor_measurements/imu" \
-"/$drone_namespace/sensor_measurements/odom" \
-"/$drone_namespace/sensor_measurements/gps" \
-"/tf" \
-"/tf_static" \
---qos-profile-overrides-path $reliability_override_file --include-hidden-topics
+cd rosbag/rosbags
 
+# Construct the rosbag record command
+rosbag_cmd="ros2 bag record"
 
+# Add topics and drone namespaces to the rosbag record command
+for drone_namespace in "${drone_namespaces[@]}"; do
+  rosbag_cmd+=" /${drone_namespace}/platform/info \
+                /${drone_namespace}/self_localization/pose \
+                /${drone_namespace}/self_localization/twist \
+                /${drone_namespace}/actuator_command/twist \
+                /${drone_namespace}/sensor_measurements/gps \
+                /${drone_namespace}/sensor_measurements/odom"
+done
 
+# Add remaining topics
+rosbag_cmd+=" /tf /tf_static --include-hidden-topics"
+
+# Execute the rosbag record command
+eval "$rosbag_cmd"
