--- rosbag/reliability_override_M200.yaml
@@ -1,38 +1,0 @@
-# reliability_override.yaml
-
-/M200/platform/info:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/self_localization/pose:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/self_localization/twist:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/actuator_command/twist:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/sensor_measurements/imu:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/sensor_measurements/odom:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/M200/sensor_measurements/gps:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/tf:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
-/tf_static:
-  reliability: best_effort
-  history: keep_all
-  depth: 5
