+++ utils/configure_onboard_computer.bash
