+++ utils/keyboard_teleop.yml
@@ -1,0 +1,16 @@
+<%
+# Input parameters
+simulation = @settings["simulation"]
+drone_namespace = @settings["drone_namespace"]
+%>
+
+attach: true
+root: ./
+startup_window: keyboard_teleop
+windows:
+  - keyboard_teleop:
+      layout:
+      panes:
+        - ros2 launch as2_keyboard_teleoperation as2_keyboard_teleoperation_launch.py
+            namespace:=<%= drone_namespace %>
+            use_sim_time:=<%= simulation %>
