+++ utils/mocap.yml
@@ -1,0 +1,10 @@
+attach: true
+root: ./
+startup_window: mocap
+windows:
+  - mocap:
+      layout:
+      panes:
+        - ros2 launch mocap_optitrack mocap.launch.py
+            namespace:=optitrack
+            config_file:=real_config/mocap.yaml
